split tests out into two folders - preview_and_dev and staging_and_prod
this enables us to take advantage of the module scoped `_driver` fixture
to make sure that accounts are logged out and cookies cleared while
moving between different accounts. To this end, the preview tests have
been split into three separate files:
* test_email_auth, for tests with the seeded email auth user
* test_seeded_user, for tests with the normal seeded user
* test_registration_and_invite, for tests with the user that is created
  every time
