+++ tests/admin/preview_and_dev/test_seeded_user.py
@@ -14,29 +14,14 @@ from tests.postman import (
 from tests.test_utils import (
     assert_notification_body,
     do_edit_and_delete_email_template,
-    do_user_can_add_reply_to_email_to_service,
-    do_user_can_invite_someone_to_notify,
-    do_user_can_update_reply_to_email_to_service,
-    do_user_registration,
     recordtime
 )
-
 
 from tests.pages import (
     DashboardPage,
     SendOneRecipient,
     SmsSenderPage,
     UploadCsvPage)
-
-
-@recordtime
-def test_registration_and_invite_flow(driver, profile, base_url):
-    do_user_registration(driver, profile, base_url)
-    do_user_can_add_reply_to_email_to_service(driver)
-    do_user_can_update_reply_to_email_to_service(driver)
-    do_user_can_update_sms_sender_of_service(driver)
-    do_user_can_add_sms_sender_to_service(driver)
-    do_user_can_invite_someone_to_notify(driver, profile, base_url)
 
 
 @recordtime
@@ -162,13 +147,6 @@ def test_send_sms_to_one_recipient(driver, profile, login_seeded_user):
     assert_dashboard_stats(dashboard_stats_before, dashboard_stats_after)
 
 
-@recordtime
-def test_email_auth(driver, profile, base_url):
-    # login email auth user
-    # assert url is research mode service's dashboard
-    pass
-
-
 @retry_on_stale_element_exception
 def get_dashboard_stats(dashboard_page, message_type, template_id):
     return {
@@ -189,44 +167,3 @@ def _get_template_count(dashboard_page, template_id):
         return 0  # template count may not exist yet if no messages sent
     else:
         return template_messages_count
-
-
-def do_user_can_update_sms_sender_of_service(driver):
-    dashboard_page = DashboardPage(driver)
-    dashboard_page.go_to_dashboard_for_service()
-
-    service_id = dashboard_page.get_service_id()
-
-    sms_sender_page = SmsSenderPage(driver)
-
-    sms_sender_page.go_to_text_message_senders(service_id)
-    sms_sender_page.click_change_link_for_first_sms_sender()
-    sms_sender_page.insert_sms_sender('first')
-    sms_sender_page.click_save_sms_sender()
-
-    main_content = sms_sender_page.get_sms_senders()
-
-    assert 'first \u2002 (default)' in main_content.text
-
-    dashboard_page.go_to_dashboard_for_service(service_id)
-
-
-def do_user_can_add_sms_sender_to_service(driver):
-    dashboard_page = DashboardPage(driver)
-    dashboard_page.go_to_dashboard_for_service()
-
-    service_id = dashboard_page.get_service_id()
-
-    sms_sender_page = SmsSenderPage(driver)
-
-    sms_sender_page.go_to_add_text_message_sender(service_id)
-    sms_sender_page.insert_sms_sender('second')
-    sms_sender_page.click_save_sms_sender()
-
-    main_content = sms_sender_page.get_sms_senders()
-
-    assert 'first \u2002 (default)' in main_content.text
-    assert 'second' in main_content.text
-    assert 'second \u2002 (default)' not in main_content.text
-
-    dashboard_page.go_to_dashboard_for_service(service_id)
